변경사항 반영 (#30)
* fix: Breed 타입 String으로 변

* Feat: 유저 탈퇴 진행중일 경우 로그인 시 401 응

* Feat: 유저 상태, 상태 업데이트 일자 컬럼 추

* Feat: 유저 탈퇴, 복구 API 구

* Test: 테스트 데이터 DATE 타입에 맞게 수정

* Chore : 펫 이미지 관련 API, 펫 정보 업데이트 API nullable하도록 수정

* Test: Pet image 테스트 코드 수정

* Chore : BreedConverter 삭제 및 pet 몸무게 실수형 변환

+++ api/src/main/java/com/petid/api/pet/dto/PetUpdateDto.java
@@ -27,7 +27,7 @@ public record PetUpdateDto() {
     }
 
     public record PetAppearanceUpdateDto(
-            int weight
+            double weight
     ){
         public PetAppearance toDomain() {
             return new PetAppearance(
